+++ sql/uninstall.php
@@ -7,6 +7,7 @@ if (!defined('_PS_VERSION_')) {
 $prefix = _DB_PREFIX_;
 
 return array(
+    "DROP TABLE IF EXISTS `{$prefix}wilden_manager_grid_preference`",
     "DROP TABLE IF EXISTS `{$prefix}wilden_manager_audit`",
     "DROP TABLE IF EXISTS `{$prefix}wilden_manager_saved_view`",
     "DROP TABLE IF EXISTS `{$prefix}wilden_manager_order_note`"